Class NativeThread: - Prepare for auto-creation of Configuration options for the thread priority of all threads by thread name... - Refactor a bit while I'm at it Class NodeStats: - Add required functions, move code from StatisticsToadlet to NodeStats Class ClientRequestScheudler / SegementedBucketChainBucketKillJob: Compile-fix: Thread priorities are integer, not short

+++ src/freenet/client/async/ClientRequestScheduler.java
@@ -880,7 +880,7 @@ public class ClientRequestScheduler implements RequestScheduler {
 		return choosenPriorityScheduler;
 	}
 
-	static final short TRIP_PENDING_PRIORITY = NativeThread.HIGH_PRIORITY-1;
+	static final int TRIP_PENDING_PRIORITY = NativeThread.HIGH_PRIORITY-1;
 	
 	public synchronized void succeeded(final BaseSendableGet succeeded, boolean persistent) {
 		if(persistent) {
